requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.16.5
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.4.0
av==15.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
colorlog==6.10.1
contourpy==1.3.0
ctranslate2==4.3.1
cycler==0.12.1
docopt==0.6.2
einops==0.8.2
exceptiongroup==1.3.1
faster-whisper==1.1.0
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
HyperPyYAML==1.2.3
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kiwisolver==1.4.7
lightning==2.6.0
lightning-utilities==0.15.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
omegaconf==2.3.0
onnxruntime==1.19.2
optuna==4.7.0
packaging==26.0
pandas==2.3.3
pillow==11.3.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.46
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.8.2
tqdm==4.67.2
transformers==4.57.6
typer==0.21.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whisperx==3.3.1
yarl==1.22.0
zipp==3.23.0
